Move Navigation component into dedicated folder
